Реализован поворот камеры с помощью сферической СК. Реализована генерация нужных объектов. Реализовано перемещение, поворот и масштабирование выбранных объектов. Реализовано изменение параметров камеры.

+++ DehaxGL/Math/Vec2i.cpp
@@ -7,5 +7,5 @@ Vec2i::Vec2i(int x, int y)
 
 long double Vec2i::length() const
 {
-    return sqrtl(x * x + y * y);
+    return std::sqrt((long double)(x * x + y * y));
 }
